Revert "Can access diffrent categories from navbar (#9)"
This reverts commit e5b5d8539d95782f6cee0187cb2dcc9ff5d47dd6.
