+++ .gitignore
@@ -5,5 +5,3 @@ frontend/my-app/package-lock.json
 rohan.pem
 backend/.env
 frontend/package-lock.json
-.vs
-backend/.vs
