+++ frontend/src/components/HomeScreen.js
@@ -1,32 +1,17 @@
-import React, { useEffect, useState } from "react";
-import "../components/products.css";
-import AllProducts from "./Products";
-import headerHome from "./images/header-home.jpg";
+import React from "react";
+import '../components/products.css';
+import AllProducts from './Products';
 
-// function to get and show all products
-export default function HomeScreen(props) {
-  const [products, setProducts] = useState([]);
 
   // function to get and show all products
-  const getProducts = async () => {
-    try {
-      const response = await fetch("http://localhost:8000/products/");
-      const jsonData = await response.json();
-      console.log(jsonData);
-      setProducts(jsonData);
-    } catch (err) {
-      console.error(err.message);
-    }
-  };
+  export default function HomeScreen() {
+  
+    return (
 
-  useEffect(() => {
-    getProducts();
-  }, []);
-
-  return (
     <div>
-      <img src={headerHome} alt="girl with mask" />
-      <AllProducts products={products} />
+    <AllProducts />
     </div>
-  );
-}
+    )
+};
+
+
