+++ frontend/src/components/ProductScreen.js
@@ -1,11 +1,19 @@
-import React from "react";
 
-import "../components/productscreen.css";
-import "../components/products.css";
-import "../index.css";
+import React from 'react';
+
+
+import '../components/productscreen.css';
+import '../components/products.css';
+import '../index.css';
+
+
+
 
 const SelectProducts = () => {
-  return <div className="main">Product Details</div>;
+ 
+  return(
+    <div className="main">Product Details</div>
+  );
 };
 
-export default SelectProducts;
+  export default SelectProducts;
