+++ frontend/src/index.css
@@ -1,12 +1,13 @@
-.main {
-  height: 551px;
+.main{
+    height: 551px;
 }
 
-.footer {
-  height: 50px;
-  width: 100%;
-  padding: 2px;
-  background-color: #1a1a1a;
-  color: whitesmoke;
-  text-align: center;
+.footer{
+    height: 50px;
+    width: 100%;
+    padding: 2px;
+    background-color:#1a1a1a;
+    color: whitesmoke;
+    text-align: center;
+    
 }
